+++ src/server/withGraphQLConfig.mjs
@@ -1,7 +1,8 @@
 /**
  * A higher-order function to decorate a Next.js custom config in
  * `next.config.js` for [`graphql-react`](https://npm.im/graphql-react), that
- * excludes server only `graphql-react/server` imports from the client bundle.
+ * excludes server only `graphql-react/server` and `next-graphql-react/server`
+ * imports from the client bundle.
  * @see [Next.js custom config docs](https://nextjs.org/docs/#custom-configuration).
  * @see [`graphql-react` `ssr` docs](https://github.com/jaydenseric/graphql-react#function-ssr).
  * @kind function
@@ -25,8 +26,15 @@ export const withGraphQLConfig = ({
 } = {}) => ({
   ...config,
   webpack(config, options) {
-    if (!options.isServer)
-      config.resolve.alias['graphql-react/server$'] = '../universal/ssrAlias'
+    if (!options.isServer) {
+      config.resolve.alias['graphql-react/server$'] =
+        '../universal/decoys/graphql-react'
+      config.resolve.alias['next-graphql-react/server/filterLinkHeader$'] =
+        '../universal/decoys/filterLinkHeader'
+      config.resolve.alias['next-graphql-react/server/mergeLinkHeaders$'] =
+        '../universal/decoys/mergeLinkHeaders'
+    }
+
     return webpack(config, options)
   }
 })