+++ src/universal/decoys/graphql-react.mjs
@@ -1,0 +1,6 @@
+/**
+ * An empty export used by [`withGraphQLConfig`]{@link withGraphQLConfig} to
+ * alias the server only module `next-graphql-react/server` for a client bundle.
+ * @ignore
+ */
+export let ssr