+++ src/universal/decoys/mergeLinkHeaders.mjs
@@ -1,0 +1,7 @@
+/**
+ * An empty export used by [`withGraphQLConfig`]{@link withGraphQLConfig} to
+ * alias the server only module `next-graphql-react/server/mergeLinkHeaders` for
+ * a client bundle.
+ * @ignore
+ */
+export let mergeLinkHeaders