+++ src/universal/ssrAlias.mjs
@@ -1,6 +1,0 @@
-/**
- * An empty named export used by [`withGraphQLConfig`]{@link withGraphQLConfig}
- * to alias the server only module `graphql-react/server` for a client bundle.
- * @ignore
- */
-export let ssr